requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.10.4
click==8.1.8
colorlog==6.9.0
dotenv==0.9.9
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
pip-chill==1.0.3
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2023.3
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
python-telegram-bot==22.0
kavenegar==1.1.2
python-dateutil==2.8.2